deleted assets folder and moved images to components
